# internal/controlpanel/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Customer control panel API for web hosting management",
        "title": "Control Panel API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/api/v1",
    "paths": {
...
        "/daemons/{id}": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Updates a daemon's configuration",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Daemons"
                ],
                "summary": "Update daemon",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Daemon ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Daemon configuration",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_edvin_controlpanel-api_internal_hosting.Daemon"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_edvin_controlpanel-api_internal_api_response.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/github_com_edvin_controlpanel-api_internal_api_response.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/github_com_edvin_controlpanel-api_internal_api_response.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Permanently removes a daemon",
                "tags": [
                    "Daemons"
                ],
                "summary": "Delete daemon",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Daemon ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_edvin_controlpanel-api_internal_api_response.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/github_com_edvin_controlpanel-api_internal_api_response.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/github_com_edvin_controlpanel-api_internal_api_response.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/daemons/{id}/disable": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Disables a daemon to stop it from running",
                "tags": [
                    "Daemons"
                ],
                "summary": "Disable daemon",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Daemon ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_edvin_controlpanel-api_internal_api_response.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/github_com_edvin_controlpanel-api_internal_api_response.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/github_com_edvin_controlpanel-api_internal_api_response.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/daemons/{id}/enable": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Enables a daemon to start running",
                "tags": [
                    "Daemons"
                ],
                "summary": "Enable daemon",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Daemon ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_edvin_controlpanel-api_internal_api_response.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/github_com_edvin_controlpanel-api_internal_api_response.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/github_com_edvin_controlpanel-api_internal_api_response.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/webroots/{id}/daemons": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns all daemons for a webroot",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Daemons"
                ],
                "summary": "List daemons",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Webroot ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "array",
                                "items": {
                                    "$ref": "#/definitions/github_com_edvin_controlpanel-api_internal_hosting.Daemon"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_edvin_controlpanel-api_internal_api_response.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/github_com_edvin_controlpanel-api_internal_api_response.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/github_com_edvin_controlpanel-api_internal_api_response.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Creates a new daemon for a webroot",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Daemons"
                ],
                "summary": "Create daemon",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Webroot ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Daemon configuration",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/github_com_edvin_controlpanel-api_internal_hosting.Daemon"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_edvin_controlpanel-api_internal_api_response.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/github_com_edvin_controlpanel-api_internal_api_response.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/github_com_edvin_controlpanel-api_internal_api_response.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "github_com_edvin_controlpanel-api_internal_api_response.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
...
        "github_com_edvin_controlpanel-api_internal_hosting.Daemon": {
            "type": "object",
            "properties": {
                "command": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "enabled": {
                    "type": "boolean"
                },
                "id": {
                    "type": "string"
                },
                "max_memory_mb": {
                    "type": "integer"
                },
                "num_procs": {
                    "type": "integer"
                },
                "proxy_path": {
                    "type": "string"
                },
                "proxy_port": {
                    "type": "integer"
                },
                "status": {
                    "type": "string"
                },
                "status_message": {
                    "type": "string"
                },
                "stop_signal": {
                    "type": "string"
                },
                "stop_wait_secs": {
                    "type": "integer"
                },
                "updated_at": {
                    "type": "string"
                },
                "webroot_id": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}